==> scrips/00_load_packages.R <==
###############################################################################
#
# Load packages
#
# For easy install, set libraries as a vector and use install.packages()
#
###############################################################################

library(biglm) 
library(boot)
library(broom)
library(data.table) 
library(DBI) 
library(dbplyr)
library(dplyr)
library(fixest)
library(formattable)
library(fs)
library(ggplot2)
library(Hmisc)
library(httr)
library(jsonlite)
library(knitr)
library(lmtest)
library(lubridate)
library(magrittr)
library(maps)
library(olsrr)
library(pbapply)
library(plm)
library(plotly)
library(quantreg)
library(RcppSimdJson)
library(readr)
library(readxl)
library(reshape)
library(rlist)
library(RSQLite)
library(sandwich)
library(sf)
library(stargazer)
library(stringr)
library(styler)
library(tidycensus)
library(tidyr)
library(tidyverse)
library(tigris)
library(tseries)
library(usmap)
library(zoo)
